Switch to SVG icons for main result mode indicator
This changes over to using SVGs for the big indicator image that shows
whether the last reading was reflection or transmission. Theme color
matching is accomplished by a search-and-replace on the loaded SVG data,
so the underlying file has to be formatted cleanly and monochrome with
elements expicitly marked as "#000000".

This makes rendering look nicer, especially on high DPI displays.