Updating and refactoring settings import/export code
Moved a lot of JSON conversion functions into the DensInterface data
structure classes.

Updated the settings export code to work with both the DPD-100 and
DPD-105.

Updated the settings import code to enforce the correct device and
handle the refactoring changes.

Still need to write the DPD-105 import code, which is likely going to
need to be a separate implementation of the import dialog.