Add ability to set a negative CAL-LO density value
There are certain lab calibration references which will have a
"negative" measured reflection density and we need to support those.

This is only going to be used for upstream calibration on the UV/VIS
densitometer, and not possible on a standalone device, so the desktop
software needs a way of explicitly setting those values.