Update Github actions to implement code signing on macOS and Windows
On macOS, initial code signing is performed on the Github
servers, but the actual notarization and stapling process is performed
separately prior to posting builds for the public.
This is because that process has the potental to introduce a days-long
delay into the build process.

On Windows, code signing uses the Microsoft Azure Trusted Signing
service, and is a fairly straightforward process. Both the app files and
the final installer are signed separately.